WIP; Build on 10.14 with target 10.7+

+++ Source/Controllers/ChannelSearchController/MIChannelSearchController.h
@@ -27,7 +27,7 @@
   NSString *currentSearchTerm;
   
   NSMutableArray *matches;
-  int currentMatch;
+  unsigned long currentMatch;
 }
 
 - (id)initWithController:(ChannelController*)controller;
